requirements.txt
singledispatchmethod
numpy>=1.14
pandas
pysam
Pillow
tqdm
tensorflow-addons
